crawler-services/parser/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.12.2
lxml==4.9.3
redis==5.0.0
